refactor: remove default impl for NlmService
